requirements.txt
absl-py==2.1.0
altgraph==0.17.4
asttokens==2.4.1
astunparse==1.6.3
atomicwrites==1.4.1
attrs==24.3.0
backports.tarfile==1.2.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
coverage==7.6.4
crypto==1.4.1
cryptography==43.0.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.7
decorator==5.1.1
docutils==0.21.2
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
huggingface-hub==0.26.0
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.7.0
keyring==25.5.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.5.0
Naked==0.1.32
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
nh3==0.2.18
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
pefile==2023.2.7
pillow==11.0.0
pkginfo==1.10.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
pure_eval==0.2.3
py==1.11.0
pyasn1==0.6.1
pycparser==2.22
pygame==2.6.1
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQtChart==5.15.7
PyQtChart-Qt5==5.15.2
pyTelegramBotAPI==4.26.0
pytest==6.2.5
pytest-pythonpath==0.7.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
readme_renderer==44.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rsa==4.9
rsa3==3.0.0.post20201106
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
shellescape==3.8.1
six==1.16.0
stack-data==0.6.3
telebot==0.0.5
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tokenizers==0.20.1
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
twine==5.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.20.2

mysql~=0.0.3
openpyxl~=3.1.5